update based on the comments
